Moved stuff from LatencyTester into P2PHandler and removed LatencyTester
P2PHandler now auto created on lobby character menu load

Switched from time since game started to real time for ping calculation

+++ Platform Fighter/Assets/_SCRIPTS/MENU/PlayerProfilePanel.cs
@@ -49,7 +49,7 @@ namespace MENU
             _playerProfiles.Add(new ProfileInfo(playerProfile, playerId));
         }
 
-        public void SetPlayerProfilePing(float ping, ulong id)
+        public void SetPlayerProfilePing(int ping, ulong id)
         {
             var profile = _playerProfiles.FirstOrDefault(x => x.PlayerId == id);
             profile.Profile.transform.Find("PingText").GetComponent<TextMeshProUGUI>().text = ping.ToString();
