+++ Platform Fighter/Assets/_SCRIPTS/NETWORKING/Events.cs
@@ -8,9 +8,7 @@ namespace NETWORKING
         public static Action<NetworkIdentity, InputChange[], Vector2, bool> OnInputsChanged;
         public static Action<NetworkIdentity> OnMatchJoined;
 
-        // Users steam id, Recievers steam id, current frame
-        public static Action<float> OnPingSent;
-        
-        public static Action<float, ulong> OnPingCalculated;
+        public static Action<int> OnPingSent;
+        public static Action<int, ulong> OnPingCalculated;
     }
 }
